# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Contacts API",
    "description": "API for managing contacts",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": ["http", "https"],
  "paths": {
    "/contacts/": {
      "get": {
        "tags": ["Contacts"],
        "description": "Get all contacts",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": ["Contacts"],
        "description": "Create a new contact",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "firstName": { "example": "Jazlyn" },
                "lastName": { "example": "Crespin" },
                "email": { "example": "jc@email.com" },
                "favoriteColor": { "example": "Purple" },
                "birthday": { "example": "2016-03-30" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Created" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "tags": ["Contacts"],
        "description": "Get a single contact by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Not Found" }
        }
      },
      "put": {
        "tags": ["Contacts"],
        "description": "Update a contact by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "firstName": { "example": "Jazlyn" },
                "lastName": { "example": "Crespin" },
                "email": { "example": "jc@email.com" },
                "favoriteColor": { "example": "Purple" },
                "birthday": { "example": "2016-03-30" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "tags": ["Contacts"],
        "description": "Delete a contact by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    }
  }
}